Force 'edit' format to html. Closes #39
The invitation token contains the user id in the format paramter for
some reason. This leads Devise to believe that setting the flash isn't
required. Had to go in a prepend_before_action becuase the
redirect-if-invitation-invalid logic was _also_ a prepend_before_action.

Also changed the custom edit action to still call `super`.