Re-worked the omniauth controller
Finally got Rubocop off my back entirely. And, of course, it's
easier to read.

The only thing that's diffent between the :twitter method and the
 :facebook is the success message. So, I wrote a generic handler that
takes that one string as an argument.

Once that was done, the `define_mthod` block was only one line and
didn't need the indirection of the `.provides_callback_for` class
method.

Finally, did some good, old-fashioned method extraction.